add error checks

+++ src/controllers/ListController.ts
@@ -7,6 +7,14 @@ export default class ListController {
     const { user } = req.headers;
     const { description } = req.body;
 
+    if (!user) {
+      return res.status(401).json({ message: "Unauthorized" });
+    }
+
+    if (!description) {
+      return res.status(400).json({ message: "Missing description" });
+    }
+
     try {
       const list = await List.create({
         user,
@@ -22,8 +30,18 @@ export default class ListController {
   getUserLists = async (req: Request, res: Response) => {
     const { user } = req.headers;
 
+    if (!user) {
+      return res.status(400).json({ message: "Missing User" });
+    }
+
     try {
       const lists = await List.find({ user });
+
+      if (!lists) {
+        return res
+          .status(404)
+          .json({ message: "There are no lists registered to this user" });
+      }
 
       return res.status(200).json(lists);
     } catch (error) {
@@ -34,9 +52,17 @@ export default class ListController {
   deleteListAndItsTasks = async (req: Request, res: Response) => {
     const { id } = req.params;
 
+    if (!id) {
+      return res.status(400).json({ message: "List id is required" });
+    }
+
     try {
-      await List.findByIdAndDelete(id);
+      const list = await List.findByIdAndDelete(id);
       await Task.deleteMany({ list: id });
+
+      if (!list) {
+        return res.status(404).json({ message: "This list doesn't exist" });
+      }
 
       return res
         .status(200)
@@ -50,12 +76,20 @@ export default class ListController {
     const { id } = req.params;
     const { description, tasks } = req.body;
 
+    if (!id || !req.body) {
+      return res.status(400).json({ message: "Missing parameters" });
+    }
+
     try {
       const list = await List.findByIdAndUpdate(
         id,
         { description, tasks },
         { new: true }
       );
+
+      if (!list) {
+        return res.status(404).json({ message: "This list doesn't exist" });
+      }
 
       return res.status(200).json(list);
     } catch (error) {
